requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
bcrypt==3.2.2
passlib[bcrypt]==1.7.4
python-jose
fastapi-mail
python-dotenv
PyJWT
mercadopago
gunicorn
httpx
python-multipart
psycopg2-binary
google-generativeai
httpx
python-multipart
sqlalchemy
PyJWT
fastapi_mail
mercadopago
google-cloud-pubsub